Create True Love
We were raised to care for one another and respect each other as well. But children don't seem to know what that means. It's  so sad that the kids have to speak mean, show ugliness and bully other children that don't have much they have or that are going through something and just need a friend to talk to. It appears some kids have learned their bulling habits from the guardians. So they may not be able to befriend others that are different. Color, size, mental and people with disabilities all should be Loved and respected as an individual. Now is  the we time we share the true meaning of caring, and respect all other children, since we are raised in a "Diverse World" But On America's money it Boldly says IN GOD WE TRUST. No matter what god you pray to. Life is too short for so much hate.